refactor: moving code around a bit, refactoring PR #12796
